Implements permissions.
If no permissions file, everybody will uses all the dropbox folder.
In the permission file, we can specify access for folders or extensions per users (DropBox display name)

Here is an example :

[
   { "folder" : "MyFolder",
     "users" : [
         "Other Person",
         "Sebastien Morel"
      ]
   },

   { "folder" : "Wherecloud",
     "users" : [
         "Sebastien Morel"
     ]
   }
]

If the permission file is on your dropbox folder that is synched, you can modify permissions on the fly and the app will react. If a user wasn't allowed and not he is, all the files for the new permissions will get downloaded and updates will be fired to the app.

If a user gets removed from a folder or extension, the cache files will get removed dans an update with the application bundle files will get fired to the app resetting the state of the resources to what has been shipped.